LayoutTests/accessibility/mac/search-subrole-expected.txt










This tests that we can detect a search field based on a few heuristics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS search.subrole is 'AXSubrole: AXSearchField'
PASS search.subrole is 'AXSubrole: AXSearchField'
PASS search.subrole is 'AXSubrole: AXSearchField'
PASS search.subrole is 'AXSubrole: AXSearchField'
PASS search.subrole is 'AXSubrole: AXSearchField'
PASS search.subrole is 'AXSubrole: AXSearchField'
PASS search.subrole is 'AXSubrole: AXSearchField'
PASS nosearch.subrole is 'AXSubrole: '
PASS nosearch.subrole is 'AXSubrole: '
PASS nosearch.subrole is 'AXSubrole: '
PASS successfullyParsed is true

TEST COMPLETE
